#954 WIP: Move the x11 code into its own header file (x11.hpp)
